Update 08 Tags and Groups.md
If the link is clicked from the page itself, it does not invoke a redirect and instead leads to a 404, whereas if the link is opened in a new tab, the redirect *does* kick in. Updated the link so that the user is navigated to the correct page straight away, removing the 404 and the need to redirect.