Clarify ramping arrival rate executor's docs
In this commit we attempt to clarify the documentation section that's
specific to the ramping arrival rate executor, as requested by #441.
We have replaced usages of the word "execution" and "executed" with
"started" to clarify the fact that the executor would not guarantee to
execute a given number of iterations per `timeUnit`, but rather would
start them.

We also expanded the provided example to illustrate its usage more
explicitly.