Fix per-VU cookie jar example
The cookie is set and overriden on the `/cookies` endpoint, and it's
sent in the GET request, but not returned in the response headers, so
it's not available in the `Response.cookies` property. Instead it's
returned in the body as JSON, which is an httpbin feature.

Previously this example would fail with:

```
ERRO[0001] TypeError: Cannot read property '0' of undefined
running at cookie has correct value (file:///tmp/test-cookies2.js:43:39(5))
default at go.k6.io/k6/js/modules/k6.(*K6).Check-fm (native)
        at file:///tmp/test-cookies2.js:42:13(39)
        at native  executor=per-vu-iterations scenario=default source=stacktrace
```