Resections and rewords create-extension docs
The "Create" docs mentioned the it's necessary to know the Go-Js bridge,
but this info is buried in the middle of the create JS tutorial.

Splitting this topic into a new "Explanation" section makes the
information easier to find, more general (it applies to output
extensions too), and it opens up a designated space for future "deep-dive"
topics.

As I extracted it, I took the time to make the tutorial a little shorter
and have a more evidently procedural format.