Move the WIP tooltip to the end of the line in row
The TableWithNestedRows adds an arrow to the beginning of the row in
the first column. One of the best practices is to move the tooltip to
the end of the line instead of at the beginning, otherwise there will
be a clash between the arrow and tooltip. This change moves all the
tooltips to the end of the line in preparation of using the nested
table.