Clarify xk6-browser running instructions
We've had reports on the forum[1] about confusions with running
xk6-browser. The main issue was that the "Build from source"
instructions build a `k6` binary, and the example mentions running
`xk6-browser`. The `InstallationInstructions` element is not flexible
enough to specify the output binary, so I chose to repeat the
instructions inline instead.  We could also parametrize this via an
`output` argument to the element, but I feel like we shouldn't
complicate it too much, as there will often be slight changes of the
instructions. It doesn't make much sense as a reusable element for me,
anyway, and the need to use HTML instead of Markdown is cumbersome
(this last point could probably be resolved with yet another Gatsby
plugin).

The other change here is to clarify the `./` and `.\` prefix required on
some systems, as that is another frequent source of confusion for users.

[1]: https://community.k6.io/t/issues-with-installing-xk6-browser-extension/4681