Remove mention of percentile query for New Relic output
As reported on the forum[1][2], our p(90) query for New Relic doesn't
produce the expected results.

After some investigation, it seems it's not possible to calculate
percentiles from metric data that our statsd output produces[3].
So instead of confusing users, it's better to delete this and clarify
the limitation (see next commit).

[1]: https://community.k6.io/t/percentiles-calculation-on-new-relic/3690

[2]: https://community.k6.io/t/how-to-get-the-correct-values-for-p-90-on-k6-newrelic/5192

[3]: https://discuss.newrelic.com/t/percentiles-of-values-from-metrics-api-with-nrql-not-working/95832