Fix orientation of 'Running k6' car icon
I'm not sure if this was caused by a font change on our side or some
Unicode update on my system, since this used to look fine a few months
ago, but on Linux and latest stable versions of Firefox and Chromium,
the car icon is rendered pointing to the right, while the "wind" icon is
pointing to the left. So it looks like the care is racing backwards :)

This still uses the same U+1F3CE glyph as before, but for some reason
renders properly with the car pointing to the left for me.

This is why it would be a good idea to abandon emojis for this use case,
and switch to SVGs (#781).