Remove benchmark results section
In order to update this information easily, we will keep benchmark
results in the k6-hardware-benchmark repository, and leave this article
for suggestions only. This way we can avoid having to continously update
this article, and it will make it more readable by being shorter.

In the future we might also remove specifically mentioning EC2, as
benchmarks will be run from GitHub Actions (Azure). Or we might want to
expand this article and give broad suggestions for the largest cloud
providers.