redo eslint configuration for examples
- Don't forbid async/await
- Don't enable everything that will work in node
- Add a bit more globals
- raise parser version to support async/await

This still means that if a syntax is not supported by the parser it will
give an error asking the user to add `requireConfigFile: false` instead
of telling them "you can't use this syntax".

Unfortunately it seems like the only way to fix that is to configure
this per each syntax which is a PITA. Hopefully k6 will soon support
even bigger parts of the syntax and that will be even less of a problem.