Add workflow for publishing technical documentation
On every merge to main, the documentation will be synced to the
`launch-k6-docs` branch in the website repository. The PR created from
that branch has a staging build that can be used to view the docs
in the context of the whole website.

On launch day, the website PR can be merged to publish the docs and
then this workflow can be updated to publish directly to the master
branch in the future.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>