Fix the browser examples which work with check
check expects a function that returns a boolean. There is enough of a
difference in what the k6 type definitions expect vs how the example
k6 browser tests work with check. If a user works with the example
browser tests as documented along with the type definitions then they
will find an error in the example code. This commit fix those examples
so that users will not see that error.