Update .lintstagedrc configuration for new repo structure
In this commit we add a rule to the lint-staged configuration file
at the root of the repository to ensure that all the markdown files
contained in the docs/sources folder are formatted using our prettier
config.

Although not absolutely necessary, we enforce prettier to use the
markdown parser, and have it write the result directly to staged files.