Improve documentation about open vs closed
The documentation on open vs closed models needed improvement. The
introduction stated:

> Different models suit different test aims

But had no discussion whatsoever of which test aims were suited for
which model.

In addition, the discussion about the drawbacks of the closed model was
incorrect. It stated:

> This effect is not ideal when the goal is to simulate a certain
> arrival rate of new VUs, or more generally throughput (e.g. requests per
> second).

The second part of this sentence should be "latency", not "throughput".
Its measuring latency that is affected by coordinated omission, as new
requests wait for old ones to finish, and that waiting time is not
included in the latency measurements. The closed model is actually
*useful* for measuring throughput.

I've:

1) added a new section "Open vs Closed Model" to the end of this
   document,

2) moved the discussion of the drawbacks of the closed model to this new
   section, as it provides the necessary context and removed some
   duplication in the text,

3) fixed the wording of the drawbacks of the closed model,

4) added some notes about when one would prefer the closed model and
   when one would prefer the open model.