Automatic update of the extension list
Automatic update of the k6 extension list in the docs/sources/next/extensions/explore.md
file in case of changes based on the https://registry.k6.io/registry.json extension registry.

The change is pushed to the extension-registry-changed branch and a pull request is opened
(or updated) about the change.