Run code blocks with latest k6 (#2027)
Run the code blocks example in documentation with the latest merge k6 instead of with latest release.

As we only run the examples that are in the next folder - they will practically always fail if they are about new functionality.

This also install chromium in the same way we run test in the k6 project.